requirements.txt
cycler==0.10.0
imageio==2.5.0
kiwisolver==1.0.1
llvmlite==0.28.0
matplotlib==3.0.3
numba==0.43.0
numpy==1.16.2
Pillow==5.4.1
pyparsing==2.3.1
python-dateutil==2.8.0
six==1.12.0
